A small benchmark program to measure parallel test execution performance
... and especially the overhead of different synchronization
approaches in JUnit's notification, to support the points in the
discussion over there at https://github.com/KentBeck/junit/pull/151.

The benchmark contains a test class with 1024 identical test methods
doing basically nothing, and this test class is run in parallel
multiple times using ParallelComputer.classes().  This way all threads
have plenty to do, while the time spent actually running the tests is
minimal compared to all the overhead spent in JUnit to invoke the
tests and collect their results.

The number of parallel runs, IOW the number of threads, is controlled
by the value of the "testclasscount" property.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>