parperf: add a simulated slow listener
JUnit's default listener is too fast to show the performance penalty
of suboptimal synchronization, because the critical sections are very
short. Therefore, add a slower listener.

The slowness is simulated by sleep() calls in the listener's
testFinished() method.  The length of the sleep() is specified by the
value of the "listenersleep" property.  The default is 0, which means
don't sleep at all.

Signed-off-by: SZEDER Gábor <szeder@fzi.de>